GhostViewer: - REFACT: big refact after many tests for synch time axis

+++ Libs/bluelab-lib/GraphTimeAxis4.h
@@ -63,7 +63,7 @@ protected:
     //
     bool mTransportIsPlaying;
     BL_FLOAT mCurrentTimeTransport;
-    long int mTransportTimeStamp;
+    double mTransportTimeStamp;
 };
 
 #endif // IGRAPHICS_NANOVG
